+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## To be Released
+- Added attack key for players to drop 'blocks'
 - Synchronized other clients facing direction
 - Broadcasts player deaths